Replace swiping over list items with seek bar
And use numbered balls instead of blocks to indicate
the weight of an argument.

Swiping over list items requires more attention and
coordination and is prone to error for some users.
Also, it's not so obvious so it's better to use a
explicit SeekBar that's always there to set the
weight.